+++ src/defaults/styles.ts
@@ -22,6 +22,7 @@ export const __colorStyles_default_preset: ColorStylesStrict = {
 		addDegrees: 0,
 		units: '',
 		extraUnits: undefined,
+		mask: 'sea',
 	},
 	custom: {
 		parent: undefined,
@@ -44,5 +45,6 @@ export const __colorStyles_default_preset: ColorStylesStrict = {
 		addDegrees: 0,
 		units: '',
 		extraUnits: undefined,
+		mask: 'sea',
 	},
 };